rewrite
Most of vproc is now rewritten to work better, when I first wrote
vproc I was pretty new to draw(2) and to plan 9 over all.

We are now able to sort the list when running the application by
right clicking anywhere to get a menu where we'll be able to modify
its behaviour, any flag that can be passed to vproc can now be
adjusted via the menu.