add functionality for sending notes to processes via middle click menu
We are now able to highlight rows which in turn indicates to the
user which process the note will be sent to. To access the menu,
middle click on it. There is no user feedback for now so if we're
trying doing something stupid we should probably do something about
it.

also we print the toolbar headers with the same function as the
rest of the list now, which makes the text line up perfectly and the
code neater.