fix: write rtime to the correct process
the `i` variable corresponds the iterator for the `/proc` directory
but that doesn't necessarily match the `proclist`, because we skip
the `/proc/trace` file, instead we use `n` to determine which index
we're writing to.

feat: refactor logic that determines if a column should be hidden

when we iterate over the headers we need to check the index value
with `skipcolumn` to determine if it should be skipped.